- Reuse same chromedriver instance for every project. - Do not convert project if it is already present on disk

+++ sb2download/ChromeDownloader/Program.cs
@@ -23,18 +23,18 @@ namespace ChromeDownloader
             }
             string outputDir = args[0];
             string[] links = File.ReadAllLines(args[1]);
+            String chromeDriverPath = Path.GetDirectoryName(Assembly.GetEntryAssembly().Location);
+            ChromeOptions options = new ChromeOptions();
+            options.AddUserProfilePreference("profile.default_content_setting_values.plugins", 1);
+            options.AddUserProfilePreference("profile.content_settings.plugin_whitelist.adobe-flash-player", 1);
+            options.AddUserProfilePreference("profile.content_settings.exceptions.plugins.*,*.per_resource.adobe-flash-player", 1);
+            // Enable Flash for this site
+            options.AddUserProfilePreference("PluginsAllowedForUrls", "https://scratch.mit.edu");
+            OpenQA.Selenium.Chrome.ChromeDriver driver = new OpenQA.Selenium.Chrome.ChromeDriver(chromeDriverPath, options);
             foreach (string line in links)
             {
                 var userProfile = Environment.GetFolderPath(Environment.SpecialFolder.UserProfile);
                 String[] files = Directory.GetFiles(userProfile, "*.sb2");
-                ChromeOptions options = new ChromeOptions();
-                options.AddUserProfilePreference("profile.default_content_setting_values.plugins", 1);
-                options.AddUserProfilePreference("profile.content_settings.plugin_whitelist.adobe-flash-player", 1);
-                options.AddUserProfilePreference("profile.content_settings.exceptions.plugins.*,*.per_resource.adobe-flash-player", 1);
-                // Enable Flash for this site
-                options.AddUserProfilePreference("PluginsAllowedForUrls", "https://scratch.mit.edu");
-                String chromeDriverPath = Path.GetDirectoryName(Assembly.GetEntryAssembly().Location);
-                OpenQA.Selenium.Chrome.ChromeDriver driver = new OpenQA.Selenium.Chrome.ChromeDriver(chromeDriverPath, options);
                 var navigate = driver.Navigate();
                 string link = line;
                 if (!link.Contains("/#editor")) {
@@ -42,35 +42,47 @@ namespace ChromeDownloader
                 }
                 Regex r= new Regex(@"/projects/(\d+)");
                 string projectNumber = r.Match(link).Groups[1].Value;
+                string targetfile = Path.Combine(outputDir, projectNumber, "index.sb2");
+                if (File.Exists(targetfile))
+                {
+                    continue;
+                }
                 navigate.GoToUrl(link);
                 System.Threading.Thread.Sleep(10000);
                 while (true)
                 {
-                    var scr1 = driver.GetScreenshot();
-                    string tempFile = Path.GetTempFileName();
-                    scr1.SaveAsFile(tempFile, ImageFormat.Png);
-                    var img = Image.FromFile(tempFile);
-                    Bitmap bitmap = new Bitmap(img);
-                    img.Dispose();
-                    File.Delete(tempFile);
-                    Color c = bitmap.GetPixel(6, 6);
-                    if (c.ToArgb() != unchecked((int)(0xff9c9ea2)))
+                    try
                     {
-                        Console.WriteLine("Waiting for load to finish...");
+                        var scr1 = driver.GetScreenshot();
+                        string tempFile = Path.GetTempFileName();
+                        scr1.SaveAsFile(tempFile, ImageFormat.Png);
+                        var img = Image.FromFile(tempFile);
+                        Bitmap bitmap = new Bitmap(img);
+                        img.Dispose();
+                        File.Delete(tempFile);
+                        Color c = bitmap.GetPixel(6, 6);
+                        if (c.ToArgb() != unchecked((int)(0xff9c9ea2)))
+                        {
+                            Console.WriteLine("Waiting for load to finish...");
+                            System.Threading.Thread.Sleep(5000);
+                            continue;
+                        }
+                        Console.WriteLine("Waiting...");
                         System.Threading.Thread.Sleep(5000);
+                        scr1.SaveAsFile("scr1.png", ImageFormat.Png);
+                        var scr2 = driver.GetScreenshot();
+                        scr2.SaveAsFile("scr2.png", ImageFormat.Png);
+                        if (scr1.AsBase64EncodedString.CompareTo(scr2.AsBase64EncodedString) == 0)
+                        {
+                            break;
+                        }
+                        File.WriteAllText("scr1.txt", scr1.AsBase64EncodedString);
+                        File.WriteAllText("scr2.txt", scr2.AsBase64EncodedString);
+                    }
+                    catch (Exception e)
+                    {
                         continue;
                     }
-                    Console.WriteLine("Waiting...");
-                    System.Threading.Thread.Sleep(5000);
-                    scr1.SaveAsFile("scr1.png", ImageFormat.Png);
-                    var scr2 = driver.GetScreenshot();
-                    scr2.SaveAsFile("scr2.png", ImageFormat.Png);
-                    if (scr1.AsBase64EncodedString.CompareTo(scr2.AsBase64EncodedString) == 0)
-                    {
-                        break;
-                    }
-                    File.WriteAllText("scr1.txt", scr1.AsBase64EncodedString);
-                    File.WriteAllText("scr2.txt", scr2.AsBase64EncodedString);
                 }
                 var scr = driver.GetScreenshot();
                 scr.SaveAsFile("final.png", ImageFormat.Png);
@@ -86,7 +98,6 @@ namespace ChromeDownloader
                 System.Threading.Thread.Sleep(10000);
                 SendKeys.SendWait(@"{Enter}");
                 System.Threading.Thread.Sleep(5000);
-                driver.Close();
                 String[] newfiles = Directory.GetFiles(userProfile, "*.sb2");
                 foreach(String file in newfiles) {
                     if (!files.Contains(file))
@@ -95,11 +106,12 @@ namespace ChromeDownloader
                         Directory.CreateDirectory(Path.Combine(outputDir, projectNumber));
                         File.Create(Path.Combine(outputDir, projectNumber, Path.GetFileNameWithoutExtension(file) + ".txt"));
                         File.WriteAllText(Path.Combine(outputDir, projectNumber, "name.txt"), Path.GetFileNameWithoutExtension(file));
-                        File.Move(file, Path.Combine(outputDir, projectNumber, "index.sb2"));
+                        File.Move(file, targetfile);
                         break;
                     }
                 }
             }
+            driver.Close();
         }
     }
 }
